move CLI11.hpp to a subfolder
It helps with the "Mark directory as excluded" in CLion.
