+++ scripts/.check-license.ignore
@@ -43,7 +43,7 @@ src/crab_utils/stats.hpp
 external/catch.hpp
 
 # Files using BSD-3-Clause
-external/CLI11.hpp
+external/CLI11/CLI11.hpp
 
 # Files using CC-BY-SA-2.5 license
 src/main/memsize_linux.hpp
