+++ src/main/run_yaml.cpp
@@ -6,7 +6,7 @@
 #include "ebpf_yaml.hpp"
 
 // Avoid affecting other headers by macros.
-#include "CLI11.hpp"
+#include "CLI11/CLI11.hpp"
 
 int main(int argc, char** argv) {
     CLI::App app{"Run YAML test cases"};
